pyproject.toml
[tool.poetry]
name = "erherium"
version = "0.1.0"
description = ""
authors = ["Anatoly <ad.uglev@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
web3 = "^6.20.2"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
certifi = "^2024.7.4"
greenlet = "^3.0.3"
fastapi-cache2 = "^0.2.2"
redis = "^5.0.8"
gunicorn = "^23.0.0"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
